chore(desktop): split move-workflow hook + WorkspaceItem below the frontend size threshold
