+++ apps/desktop/src/components/workspace/shell/sidebar/WorkspaceItem.tsx
@@ -33,6 +33,7 @@ import {
   SidebarStatusIndicatorView,
 } from "./SidebarIndicators";
 import { SidebarWorkspaceGitGlyph } from "./SidebarWorkspaceGitGlyph";
+import { WorkspaceDeleteConfirmMenu } from "./WorkspaceDeleteConfirmMenu";
 import { WorkspaceItemMenu } from "./WorkspaceItemMenu";
 import { WorkspaceRenamePopover } from "./WorkspaceRenamePopover";
 import { ProductSidebarWorkspaceRow } from "@proliferate/product-ui/sidebar/ProductSidebarRepositories";
@@ -257,36 +258,10 @@ export function WorkspaceItem({
       {(close) => (
         <>
           {doneConfirmOpen ? (
-            <>
-              <div className="px-2.5 py-2 text-sm text-foreground">
-                <div className="font-medium">Delete workspace?</div>
-                <div className="mt-1 text-xs leading-4 text-muted-foreground">
-                  This removes the local worktree, workspace record, chat history, and local agent
-                  artifacts for this workspace. Commits, branches, and pull requests are not deleted.
-                </div>
-                <div className="mt-1 text-xs leading-4 text-muted-foreground">
-                  This cannot be undone from Proliferate.
-                </div>
-              </div>
-              <PopoverMenuItem
-                icon={<Trash className="size-3.5 shrink-0 text-muted-foreground" />}
-                label="Delete workspace"
-                variant="sidebar"
-                onClick={() => {
-                  close();
-                  setDoneConfirmOpen(false);
-                  onMarkDone?.();
-                }}
-              />
-              <PopoverMenuItem
-                label="Cancel"
-                variant="sidebar"
-                onClick={() => {
-                  close();
-                  setDoneConfirmOpen(false);
-                }}
-              />
-            </>
+            <WorkspaceDeleteConfirmMenu
+              onConfirm={() => { close(); setDoneConfirmOpen(false); onMarkDone?.(); }}
+              onCancel={() => { close(); setDoneConfirmOpen(false); }}
+            />
           ) : (
             <>
               {onRename && (
